requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
chilkat==10.1.1
idna==3.10
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
starkbank-ecdsa==2.2.0
tzdata==2024.2
urllib3==2.3.0
